# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: >
    API for the WASAText messaging application, which allows users to send and receive
    individual or group messages with support for text, photos, and reactions.
    Includes simplified login, user management, conversations, messages, and group handling.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "user-id"
  schemas:
    ErrorResponse:
      type: object
      description: Standard error response payload.
      properties:
        error:
          type: string
          description: Human-readable error message describing what went wrong.
          example: "An error occurred."
          minLength: 1
          maxLength: 255
          pattern: '^.+$'
      required:
        - error
    User:
      type: object
      description: Represents an application user.
      properties:
        id:
          type: string
          description: Unique user identifier.
          example: "user123"
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        name:
          type: string
          description: Display name of the user.
          example: "Maria"
          minLength: 1
          maxLength: 50
          pattern: '^[A-Za-zÀ-ÿ0-9 .''-]{1,50}$'
        photo:
          type: string
          format: uri
          description: Public URL of the user profile photo.
          example: "https://example.com/maria.jpg"
          minLength: 1
          maxLength: 2048
          pattern: '^https?://.+$'
      required: [name]
    Conversation:
      type: object
      description: Represents a conversation (direct or group).
      properties:
        id:
          type: string
          description: Unique conversation identifier.
          example: "conv123"
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        type:
          type: string
          description: Type of conversation (direct or group).
          enum: [direct, group]
          example: "group"
          minLength: 5
          maxLength: 6
          pattern: '^(direct|group)$'
        name:
          type: string
          description: Human-readable name of the conversation (mainly for groups).
          example: "Work Friends"
          minLength: 1
          maxLength: 100
          pattern: '^[A-Za-zÀ-ÿ0-9 .''-]{1,100}$'
        photo:
          type: string
          format: uri
          description: Public URL of the conversation or group photo.
          example: "https://example.com/group-photo.jpg"
          minLength: 1
          maxLength: 2048
          pattern: '^https?://.+$'
        lastMessagePreview:
          type: string
          description: Short preview of the most recent message in the conversation.
          example: "See you tomorrow?"
          minLength: 0
          maxLength: 140
          pattern: '^[\s\S]*$'
        lastMessageAt:
          type: string
          format: date-time
          description: Timestamp of the most recent message in the conversation (ISO 8601).
          example: "2025-11-04T20:00:00Z"
        lastMessageSenderId:
          type: string
          description: ID of the user who sent the last message.
          example: "user123"
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        lastMessageSenderName:
          type: string
          description: Name of the user who sent the last message.
          example: "Juan"
          minLength: 1
          maxLength: 50
          pattern: '^[A-Za-zÀ-ÿ0-9 .''-]{1,50}$'
        lastMessageStatus:
          type: string
          description: Status of the last message.
          enum: [delivered, read]
          example: "read"
          minLength: 4
          maxLength: 9
          pattern: '^(delivered|read)$'
        unreadCount:
          type: integer
          description: Number of unread messages for the requesting user.
          example: 5
          minimum: 0
        participants:
          type: array
          description: List of users participating in the conversation.
          minItems: 1
          maxItems: 512
          items:
            $ref: "#/components/schemas/User"
      required: [id, type, participants]
...
  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      description: Conversation ID.
      schema:
        type: string
        example: "conv123"
        minLength: 1
        maxLength: 64
        pattern: '^[A-Za-z0-9_-]+$'
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /conversations/{conversationId}/users/me:
    delete:
      tags:
        - Groups
      summary: Leave group
      description: >
        Remove the authenticated user from the specified group conversation.
        **Restriction:** This operation fails (403) if attempted on a 'direct' (private) conversation.
      operationId: leaveGroup
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        '204':
          description: User left the group successfully.
        '400':
          description: Missing conversation ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          description: Forbidden. The conversation is not a group, or the user is not a participant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Conversation not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /conversations/{conversationId}/name:
    put:
      tags:
        - Groups
      summary: Change group name
      description: >
        Update the display name of a group conversation.
        **Restriction:** Only works for group conversations (returns error 500 if attempted on private chats).
      operationId: setGroupName
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload containing the new group name.
              properties:
                name:
                  type: string
                  description: New name for the group.
                  example: "New Group Name"
                  minLength: 1
                  maxLength: 100
              required:
                - name
      responses:
        '200':
          description: Group name changed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        '400':
          description: Invalid name length.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          description: User is not a participant of the group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Conversation not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error (or attempt to rename private chat).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /conversations/{conversationId}/photo:
    put:
      tags:
        - Groups
      summary: Update group photo
      description: >
        Upload and set a new photo for a group conversation.
        **Restriction:** Only works for group conversations (returns error 500 if attempted on private chats).
      operationId: setGroupPhoto
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Multipart form data containing the new group photo.
              properties:
                photoFile:
                  type: string
                  format: binary
                  description: Binary image file.
              required:
                - photoFile
      responses:
        '200':
          description: Group photo updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        '400':
          description: Missing file or invalid form data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          description: User is not a participant of the group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Conversation not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error (or attempt to update private chat).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...